Refactor the repos module into multiple files
